+++ src/frappe/components/Icons/MaximizeIcon.vue
