+++ src/frappe/components/Icons/MinimizeIcon.vue
