+++ src/frappe/components/Icons/StepsIcon.vue
