+++ src/frappe/components/OnboardingSteps.vue
@@ -57,7 +57,7 @@
   </div>
 </template>
 <script setup>
-import { useOnboarding } from '../../composables/onboarding'
+import { useOnboarding } from '../composables/onboarding'
 
 const props = defineProps({
   appName: {
